hud: move editstatusline into its own file
